refactor(engine): centralize EXPLORA flags and silence clippy
Add engine/src/flags.rs as the single registry for every EXPLORA_* process
flag. Play flags stay active in every build; DEBUG_ONLY flags compile out of
release and dist so playable builds always run Cinematic with no quality
knobs. Rewire quality, audio, HUD, wind, spawn, FX, pacing, freeze, RT
shadows, mesh shaders, IBL, burst, present, and GPU diagnostic call sites
through the new accessors. Update README, pgo-engine.sh, and
benchmark-frames.py for the Cinematic-only contract.

Also clear workspace clippy and build-script warnings: mark pointer-taking
FFI exports unsafe with safety docs, replace manual modulo/chunks/clamp
patterns, add Default for Effects and TrailPool, allow fixed-arity Vulkan
entry points, and stop emitting baker stats as cargo:warning.

+++ crates/airframe/src/util.rs
@@ -87,7 +87,7 @@ pub fn f32_to_f16(v: f32) -> u16 {
         let half = (shifted + 0xfff + ((shifted >> 13) & 1)) >> 13;
         return (sign | half) as u16;
     }
-    let half = ((mant + 0xfff + ((mant >> 13) & 1)) >> 13) as u32;
+    let half = (mant + 0xfff + ((mant >> 13) & 1)) >> 13;
     (sign | ((exp as u32) << 10) | half.min(0x3ff)) as u16
 }
